\name{readwinch}
\alias{readwinch}
\title{ Read WIN channel file }
\description{
  Read WIN channel file
}
\usage{
readwinch(fn, wigdir = "/home/lees/Site/Fuji/WIG")
}
\arguments{
  \item{fn}{file name}
  \item{wigdir}{name of directory where waveforms reside}
}

\value{
 
  
}
\references{}
\author{Jonathan M. Lees<jonathan.lees@unc.edu>}
\note{
}
\seealso{}
\examples{}
\keyword{misc}
