\name{WINRPACK-package}
\alias{WINRPACK-package}
\alias{WINRPACK}
\docType{package}
\title{
Read and convert WIN format to R
}
\description{
Package reads in WIN format data from Japan and converts to R
}
\details{
\tabular{ll}{
Package: \tab WINRPACK\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2008-03-24\cr
License: \tab GPL\cr
}


}
\author{
Jonathan M. Lees<jonathan.lees\@unc.edu>
Maintainer:Jonathan M. Lees<jonathan.lees\@unc.edu>

}
\references{

}

\keyword{ package }
\seealso{
RSEIS, Rsac
}
\examples{

}
