\name{XWINdata}
\alias{XWINdata}
\title{Extract WIN format}
\description{
  Extract WIN format
}
\usage{
XWINdata(fn, stasel = stasel, PLOT = FALSE)
}
\arguments{
   \item{fn}{file name}
  \item{stasel}{selection of stations}
  \item{PLOT}{logical wheaterh to plot}
 
}

\value{
 List of stuff
  
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\keyword{misc}
