\name{XWINdataASK}
\alias{XWINdataASK}
\title{extract win format to ASCII}
\description{
  extract win format to ASCII
}
\usage{
XWINdataASK(fn, stasel = stasel, PLOT = FALSE)
}
\arguments{
  \item{fn}{file name}
  \item{stasel}{selection of stations}
  \item{PLOT}{logical wheaterh to plot}
}
\details{
}
\value{
  \item{List}{stuff}
  
}
\references{}
\author{Jonathan M. Lees<jonathan.lees@unc.edu>}
\note{
}
\seealso{}
\examples{}
\keyword{misc}
