\name{XWINdata}
\alias{XWINdata}
\title{Extract WIN format}
\description{
  Extract WIN format
}
\usage{
XWINdata(fn, stasel = stasel, PLOT = FALSE)
}
\arguments{
   \item{fn}{file name}
  \item{stasel}{selection of stations}
  \item{PLOT}{logical wheaterh to plot}
 
}
\details{
}
\value{
 List of stuff
  
}
\references{}
\author{Jonathan M. Lees<jonathan.lees@unc.edu>}
\note{
}
\seealso{}
\examples{}
\keyword{misc}
