\name{XWINdataASK}
\alias{XWINdataASK}
\title{Extract win format to ASCII}
\description{
  Extract win format to ASCII
}
\usage{
XWINdataASK(fn, stasel = stasel, PLOT = FALSE)
}
\arguments{
  \item{fn}{file name}
  \item{stasel}{selection of stations}
  \item{PLOT}{logical wheaterh to plot}
}

\value{
  \item{List}{stuff}
  
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\keyword{misc}
